app/requirements.txt
fastapi
python-multipart
PyMySQL
boto3
kafka-python